requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.29.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

